# examples/java/prophet_example_spring/gen/spring-boot/src/main/java/com/example/prophet/commerce_local/generated/api/filters/OrderStateFilter.java
package com.example.prophet.commerce_local.generated.api.filters;

import javax.annotation.processing.Generated;
import com.example.prophet.commerce_local.generated.domain.OrderState;
import java.util.List;

@Generated("prophet-cli")
public record OrderStateFilter(
    OrderState eq,
    List<OrderState> in
) {

    public static Builder builder() {
        return new Builder();
    }

    public static final class Builder {
        private OrderState eq;
        private List<OrderState> in;

        public Builder eq(OrderState value) {
            this.eq = value;
            return this;
        }

        public Builder in(List<OrderState> value) {
            this.in = value;
            return this;
        }
        public OrderStateFilter build() {
            return new OrderStateFilter(
                eq,
                in
            );
        }
    }
}