# examples/node/prophet_example_express_mongoose/gen/node-express/src/generated/domain.ts
// Code generated by prophet-cli. DO NOT EDIT.

export type Money = number;

export interface OrderRef {
  orderId: string;
}

export interface UserRef {
  userId: string;
}

export interface Address {
  line1: string;
  city: string;
  countryCode: string;
}

export interface ApprovalContext {
  approver: UserRef;
  watchers?: UserRef[];
  reason?: string;
}

export type OrderState = "created" | "approved" | "shipped";

export interface Order {
  orderId: string;
  customer: UserRef;
  totalAmount: number;
  discountCode?: string;
  tags?: string[];
  shippingAddress?: Address;
  state: OrderState;
}

export interface User {
  userId: string;
  email: string;
}