# examples/python/prophet_example_flask_sqlalchemy/gen/python/src/generated/domain.py
# Code generated by prophet-cli. DO NOT EDIT.
from __future__ import annotations

from dataclasses import dataclass
from typing import Any, Dict, List, Literal, Optional

Money = float

@dataclass(kw_only=True)
class OrderRef:
    orderId: str

@dataclass(kw_only=True)
class UserRef:
    userId: str

@dataclass(kw_only=True)
class Address:
    line1: str
    city: str
    countryCode: str

@dataclass(kw_only=True)
class ApprovalContext:
    approver: UserRef
    watchers: Optional[List[UserRef]] = None
    reason: Optional[str] = None

OrderState = Literal['created', 'approved', 'shipped']

@dataclass(kw_only=True)
class Order:
    orderId: str
    customer: UserRef
    totalAmount: float
    discountCode: Optional[str] = None
    tags: Optional[List[str]] = None
    shippingAddress: Optional[Address] = None
    state: OrderState

@dataclass(kw_only=True)
class User:
    userId: str
    email: str